# docs/openapi.yaml
{
    "openapi": "3.0.2",
    "info": {
        "title": "ugc",
        "description": "User movie progress view",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/view_progress": {
            "post": {
                "tags": [
                    "view"
                ],
                "summary": "Message about watching the film",
                "description": "Message about watching the film",
                "operationId": "view_progress_api_v1_view_progress_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ViewProgressRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/like": {
            "post": {
                "tags": [
                    "likes"
                ],
                "summary": "Like",
                "description": "Add like",
                "operationId": "like_api_v1_like_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LikeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dislike": {
            "post": {
                "tags": [
                    "likes"
                ],
                "summary": "Dislike",
                "description": "Add dislike",
                "operationId": "dislike_api_v1_dislike_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DislikeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "DislikeRequest": {
                "title": "DislikeRequest",
                "required": [
                    "movie_id",
                    "user_id"
                ],
                "type": "object",
                "properties": {
                    "_id": {
                        "title": " Id",
                        "type": "string",
                        "format": "uuid",
                        "default": "b17d77a8-7814-4bc6-bb84-bca28a6d9f62"
                    },
                    "movie_id": {
                        "title": "Movie Id",
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "title": "User Id",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "description": "Dislike request."
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "LikeRequest": {
                "title": "LikeRequest",
                "required": [
                    "movie_id",
                    "user_id"
                ],
                "type": "object",
                "properties": {
                    "_id": {
                        "title": " Id",
                        "type": "string",
                        "format": "uuid",
                        "default": "b17d77a8-7814-4bc6-bb84-bca28a6d9f62"
                    },
                    "movie_id": {
                        "title": "Movie Id",
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "title": "User Id",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "description": "Like request."
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            },
            "ViewProgressRequest": {
                "title": "ViewProgressRequest",
                "required": [
                    "movie_id",
                    "user_id",
                    "viewed_frame"
                ],
                "type": "object",
                "properties": {
                    "movie_id": {
                        "title": "Movie Id",
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "title": "User Id",
                        "type": "string",
                        "format": "uuid"
                    },
                    "viewed_frame": {
                        "title": "Viewed Frame",
                        "type": "integer"
                    }
                },
                "description": "View progress request body."
            }
        }
    }
}